refactor(geode-core): group eigensolvers under eigen [5/10]
Child [5/10] of the geode-core hierarchical-namespace epic #377,
mirroring merged siblings #378/#379/#380/#381 and PR #391.

Pure refactor — no behavior change. `git mv` R100 renames preserve
blame; all `#[cfg(...)]` gates and `pub(crate)` visibilities are kept.

Layout (directory-backed `eigen/`):
- eigen/dense.rs    (git mv eigen.rs)         -> eigen::dense
- eigen/lanczos.rs  (git mv lanczos.rs)       -> eigen::lanczos
- eigen/complex/    directory (curator's recommended layout to avoid
  colliding the two `#[cfg(test)] mod tests`):
    - complex/dense.rs   (git mv complex_eigen.rs)   private leaf
    - complex/lanczos.rs (git mv complex_lanczos.rs) private leaf
    - complex/mod.rs     re-exports one `eigen::complex` namespace;
      preserves `pub(crate)` spmv/solve_with_lu, leaf-private spmv_add
- eigen/arpack.rs   (git mv arpack.rs)        -> eigen::arpack
  (kept `#[cfg(feature = "arpack")]` + `#![allow(unsafe_code)]`)
- eigen/mod.rs, eigen/complex/mod.rs: new group `//!` rustdoc

lib.rs: drop the 5 top-level module decls, add `pub mod eigen;`.
The 5 `pub use` blocks become `#[deprecated]` flat-root item shims at
canonical `eigen::{dense,lanczos,complex,arpack}` paths; the arpack
shim keeps `#[cfg(feature = "arpack")]`.

prelude.rs: add high-traffic eigen items from canonical paths.

Same-PR caller migration (lib `src/` target only): Class A imports +
Class B intra-doc links repointed to the new paths. Flat-root item
callers in tests/benches/examples stay green via the deprecated shims
(deferred to cleanup child #10).

Closes #382
