+++ crates/geode-core/src/eigen/dense.rs
