+++ crates/geode-core/src/prelude.rs
@@ -1,7 +1,7 @@
 //! Common imports for geode-core: `use geode_core::prelude::*;`.
 //!
 //! Re-exports from the canonical module paths (`crate::assembly`,
-//! `crate::backend`, `crate::derham`, `crate::elements`,
+//! `crate::backend`, `crate::derham`, `crate::eigen`, `crate::elements`,
 //! `crate::solver`, `crate::traits`) — never the deprecated root shims — so
 //! glob-importing the prelude stays warning-free under `-D warnings`.
 //! Later children of the namespace reorg add their own groups here.
@@ -11,6 +11,14 @@ pub use crate::assembly::p1::{GlobalSystem, assemble_global_p1};
 pub use crate::assembly::sparse::{SparseSystem, global_system_to_sparse};
 pub use crate::backend::{DefaultBackend, DeviceInfo, device_info, smoke_add};
 pub use crate::derham::{curl_map, divergence_map, gradient_map};
+pub use crate::eigen::complex::{
+    ComplexEigenSolver, FaerComplexEigensolver, SparseComplexEigenSolver,
+    SparseComplexShiftInvertLanczos,
+};
+pub use crate::eigen::dense::{
+    EigenSolver, FaerDenseEigensolver, apply_dirichlet_bc, burn_matrix_to_faer, cube_interior_mask,
+};
+pub use crate::eigen::lanczos::{SparseEigenSolver, SparseShiftInvertLanczos};
 pub use crate::elements::nedelec::{
     NedelecLocalMatrices, batched_nedelec_local_matrices, tet_edges,
 };
